// src/payload-types.ts
export interface Fruit {
  id: string;
  blocks?:
    | (
        | {
            AppleKind?: string | null;
            showConditionalApple?: boolean | null;
            conditionalFields?: {
              ConditionalFieldApple?: string | null;
            };
            seedApple?: (string | Seed)[] | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'apples';
          }
        | {
            PearType?: string | null;
            showConditionalPear?: boolean | null;
            ConditionalFieldPear?: string | null;
            seedPear?: (string | Seed)[] | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'pears';
          }
      )[]
    | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "seeds".
 */
export interface Seed {
  id: string;
  name?: string | null;
  updatedAt: string;
  createdAt: string;
}